+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Notable changes to this project goes here
 
 ### Added
 
-- Support for disabling HTTP method restrictions when registering routes.
+- Ability to disable HTTP method restrictions when registering routes.
+- New special fallback method `*` to match any method during lookups.
 
 ## [v3.4.3] - 26.09.2025